azure-functions/DeviceRealtimeAPI/requirements.txt
azure-functions>=1.18.0
azure-cosmos>=4.5.0
